Fixed AttributeError when using raw SQL
When executing raw SQL in my application, the report generation would
fail. This was because the default report template uses `text.params`;
this attribute does not exist if the `text` passed to `QueryStats` is a
plain string (which it is when doing raw queries).

This fix introduces the `params` attribute for `QueryStats` objects
which is an empty list if the `text` passed in is a string.